backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
playwright==1.47.2
